% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_test.R
\name{pathway_analysis}
\alias{pathway_analysis}
\title{hypergeometric test for RNA-seq}
\usage{
pathway_analysis(sig.genes, class = c("mmu", "hsa", "gga"), mc.cores = 8)
}
\arguments{
\item{sig.genes}{A vector of significantly changed genes}

\item{class}{Specify which type of gene}

\item{mc.cores}{Specify how many cores to use}
}
\value{
A list of all significantly changed pathways, and all pathways with p value
}
\description{
hypergeometric test for RNA-seq
}
\examples{
deg = grab_deg_from_cuffdiff(gene_exp.diff)
sig.genes = deg$external_gene_name
path_res = pathway_analysis(sig.genes, class = "mmu")
}
